+++ src/BDSComponentFactory.cc
@@ -2358,6 +2358,20 @@ void BDSComponentFactory::CalculateAngleAndFieldSBend(Element const* el,
       angle = el->angle * CLHEP::rad;
       field = FieldFromAngle(angle, arcLength);
     }
+  // un-split sbends are effectively rbends - don't construct a >pi/2 degree rbend
+  // also cannot construct sbends with angles > pi/2
+  if (BDSGlobalConstants::Instance()->DontSplitSBends() && (std::abs(angle) > CLHEP::pi/2.0))
+    {
+      G4cerr << "Error: the unsplit sbend "<< el->name << " cannot be constucted as it's bending angle "
+             << "is defined to be greater than pi/2." << G4endl;
+      exit(1);
+    }
+  else if (std::abs(angle) > CLHEP::pi*2.0)
+    {
+      G4cerr << "Error: the sbend "<< el->name << " cannot be constucted as it's bending angle "
+             << "is defined to be greater than 2 pi." << G4endl;
+      exit(1);
+    }
 }
 
 void BDSComponentFactory::CalculateAngleAndFieldRBend(const Element* el,
@@ -2390,6 +2404,12 @@ void BDSComponentFactory::CalculateAngleAndFieldRBend(const Element* el,
       field = el->B * CLHEP::tesla;
       G4double bendingRadius = brho / field; // in mm as brho already in g4 units
       angle = 2.0*std::asin(chordLength*0.5 / bendingRadius);
+      if (std::isnan(angle))
+        {
+          G4cerr << "Field too strong for element " << el->name << ", magnet bending angle will be greater than pi." << G4endl;
+          exit(1);
+        }
+
       arcLengthLocal = bendingRadius * angle;
     }
   else
@@ -2410,6 +2430,13 @@ void BDSComponentFactory::CalculateAngleAndFieldRBend(const Element* el,
 
   // Ensure positive length despite sign of angle.
   arcLength = std::abs(arcLengthLocal);
+
+  if (std::abs(angle) > CLHEP::pi/2.0)
+    {
+      G4cerr << "Error: the rbend "<< el->name << " cannot be constucted as it's bending angle "
+             << "is defined to be greater than pi/2." << G4endl;
+      exit(1);
+    }
 }
 
 G4double BDSComponentFactory::BendAngle(const Element* el) const